backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
black==25.11.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.1.8
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
filetype==1.2.0
firecrawl-py==4.19.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.0.2
olefile==0.47
packaging==26.0
pathspec==1.0.4
pgvector==0.4.2
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.4.1
psutil==7.2.2
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pypdf==6.9.1
pypdfium2==5.6.0
pytest==8.4.2
python-dotenv==1.2.1
python-iso639==2025.2.18
python-magic==0.4.27
pytokens==0.4.1
RapidFuzz==3.13.0
regex==2026.1.15
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
tomli==2.4.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.6
wrapt==2.1.2
yarl==1.22.0
zipp==3.23.0
